requirements.txt
asttokens==2.1.0
backcall==0.2.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.4.5
jupyter_core==5.0.0
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
stack-data==0.6.1
statsmodels==0.13.5
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.5.0
wcwidth==0.2.5
xgboost==1.7.1
